+++ modules/tinymce/src/plugins/image/main/ts/ui/MainTab.ts
@@ -71,6 +71,8 @@ const makeItems = (info: ImageDialogInfo) => {
     ]
   };
 
+  const dialogType = (hasColumns: boolean) => hasColumns ? { type: 'grid', columns: 2 } : { type: 'panel' };
+
   return Arr.flatten<any>([
     [ imageUrl ],
     imageList.toArray(),
@@ -79,8 +81,7 @@ const makeItems = (info: ImageDialogInfo) => {
     info.hasImageTitle ? [ imageTitle ] : [],
     info.hasDimensions ? [ imageDimensions ] : [],
     [{
-      type: 'grid',
-      columns: 2,
+      ...dialogType(info.hasImageCaption),
       items: Arr.flatten([
         classList.toArray(),
         info.hasImageCaption ? [ caption ] : []